refactor(trie): move repair-trie verifier to trie-parallel
Amp-Thread-ID: https://ampcode.com/threads/T-019dba25-4d05-7662-a03e-3a51d7b5cf21
