+++ crates/trie/parallel/Cargo.toml
@@ -13,14 +13,18 @@ workspace = true
 
 [dependencies]
 # reth
+reth-db-api.workspace = true
 reth-primitives-traits = { workspace = true, features = ["dashmap", "std"] }
 reth-execution-errors.workspace = true
 reth-provider.workspace = true
+reth-storage-api = { workspace = true, features = ["db-api"] }
 reth-storage-errors.workspace = true
 reth-tasks = { workspace = true, features = ["rayon"] }
 reth-trie.workspace = true
+reth-trie-db.workspace = true
 
 # alloy
+alloy-consensus.workspace = true
 alloy-eip7928.workspace = true
 alloy-evm.workspace = true
 alloy-primitives.workspace = true
@@ -76,4 +80,5 @@ test-utils = [
     "reth-trie/test-utils",
     "reth-tasks/test-utils",
     "reth-trie-sparse?/test-utils",
+    "reth-db-api/test-utils",
 ]
